add fedup.download.setup_media_mount
This function takes a FstabEntry object (as returned by
fedup.media.mounts() etc.) and writes out a systemd .mount unit in the
system-upgrade.target.wants directory.

This will make the system mount the install media to $UPGRADELINK/media
before the upgrade starts, so we can install packages from it. Hooray!