Add fedup.conf module, write packagedir/upgrade.conf
There might be other things we want to clean up / refer to in the
future, so it might be better if we had a structured file that contained
info about the upgrade.

Currently we write two sections, like this:

  [cleanup]
  dirs = <dirs to clean up>

  [boot]
  kernel = <path to upgrader kernel>
  initrd = <path to upgrader initrd>

This allows other tools to clean up after us or examine the state of the
upgrade or whatever.