remove grubby, let new-kernel-pkg set default
The whole reason for switching to new-kernel-pkg was that grubby fails
when it doesn't find the right config. So why did I think using grubby
to get/set the default kernel would work?

Let new-kernel-pkg set the defaults. Anything that goes wrong here is
new-kernel-pkg's fault.

Also, fix the --make-default flag and make it the default.