Fix loopmount handling (and fedup --iso)
A mount entry for a loop-mounted filesystem will have dev="/dev/loopX",
but that's not so useful for write_systemd_unit.

fix_loop_entry() rewrites a FstabEntry so:

  a) dev is the backing file for the loop device, and
  b) "loop" is added to the mount options.

So if we're writing a .mount unit for a loop device, fix it up like this
and it'll actually work.

This patch also fixes loopmount() so it will accept a mntpoint argument,
creating a temporary directory if none is provided.