Fix textoutput traceback on weird consoles (#888468)
Some consoles (e.g. the HVC on POWER systems) will return (0, 0) for
`stty size` or TIOCGWINSZ.

If that happens, we'll end up calculating the barwidth as some negative
number, which leads to a ValueError traceback when we try to use it as a
format width.

So: just assume that dumb terminals are 80 columns wide (since they
usually are).