cli: better error messages for --iso problems
Give better error messages for the two most obvious problems we have
with --iso: file not found, or the file is on a removable device.

(We don't know how to recursively set up mounts for ISOs on removable
devices, so we need to reject this.. for now.)

Also, remove the '[TODO]' from the help message, 'cuz this works now.
Yay!