Put SELinux in permissive mode during upgrade (#896010)
There are some places where the old SELinux policy and the new SELinux
policy don't match. So there's some stuff that the new policy will deny
that the old policy allowed.

Like maybe something in your system startup used to write to /tmp, but
in the new system it writes a tempfile to a private directory.

So: when we switch-root to the old system to start it up, part of the
startup might get denied, and the system might not start up properly,
and the upgrade might not start.

If we had a way to make SELinux use the old policy in the old system and
new policy in the new system, that would probably be a better fix, but
for now...