fedup.sysprep: make /lib/modules/$(uname -r) before reboot
I REMEMBERED WHAT KERNELVER WAS FOR!! :D (see commit a2befeaa)

Creating this directory before rebooting will mean that we don't have to
play silly games trying to remount / read-write and junk to create it
during initramfs. Sometimes that causes weird problems, so it's better
to avoid it if possible.