cache initrd so we don't have to refetch if it's modified
If we append stuff to the initrd (e.g. because we're applying updates or
inserting mdadm.conf - see sysprep.prep_boot()), its checksum will
change, and if you re-run fedup we'll think it's corrupt or something
and download it all over again.

So to save us a little bandwidth and time, save the original to the
cachedir, then copy it to initrdpath (where we can modify it freely).

This will make us use an extra ~25MB in cachedir, but that's like a 2%
increase. And disk is less precious than bandwidth or time.