Better error messages on TransactionError (#896144)
So: if we're actually out of diskspace, the big damn list of RPM
problems is pretty useless. It also makes the original "Transaction
failed" message scroll offscreen, so people miss that key bit of info.

Instead, we can slice-and-dice the list into a short, useful list of
"Need XXgb on device XX" errors - and then put an extra "ERROR" message
at the end for good measure. Hooray!