fedup.py: error message when there's no updates
It's possible that you'll end up with no available repos (except your
custom instrepo, which has no packages), and then you'd get the message:

  Your system is already upgraded!

and we'll exit with returncode 0 (i.e. success). Which this isn't.

Head that condition off a little early and print an appropriate message
and return failure instead.