fedup.spec: add "Obsoletes: preupgrade" (#874495)
PackageKit in F19 and later drops the requirement on preupgrade, and
AFAIK it was the only thing in Fedora that required it.

In that case, we can safely Obsolete preupgrade, thus removing it from
people's systems without breaking anything. ATONEMENT AT LAST.

Big ups to kalev for doing the legwork/testing here. You may exchange
a copy/image of this git commit for one (1) free beer or other beverage
of your choice.