upgrade-prep.sh: don't trigger autofs mounts
upgrade-prep.sh loops through all the mounts in /proc/mounts and tries
to unmount anything that has a file named "/.please-unmount".

But: if we try to stat "/.please-unmount" on an autofs mount, it'll
cause the mount to get mounted, which is not what we want at all.

So: skip anything with fstype "autofs" in that loop.