Use system-upgrade-generator instead of 'systemd.unit=...'
Rather than relying on "systemd.unit=system-upgrade.target" to start the
upgrade, we should follow the precedent set by:

  http://freedesktop.org/wiki/Software/systemd/SystemUpdates/

and use a systemd generator to start the upgrade if the right conditions
are met.

In the case of system upgrades, our generator will switch the default
target to "system-upgrade.target" iff /system-upgrade/.start exists.

Once that happens, the first thing that the upgrade tool should do is
remove /system-upgrade/.start, to avoid boot loops.