fedup.download: check package signatures (#877623)
This makes fedup verify the signatures of downloaded packages.
(Only if the corresponding .repo file has 'gpgcheck=1', of course).

The signature check works as usual: if the key used to sign the package
is already trusted, that key is used to verify the package.

If the package is signed by an *untrusted* key, the 'gpgkey=URL' item
in the corresponding .repo file will be examined; any yum variables in
the listed URL will be replaced as usual. ($releasever will be set to
the VERSION given with '--network'.)

If the resulting key file:

  a) has the correct key ID, *and*
  b) belongs to a package that was signed with a trusted key, *and*
  c) the file hasn't been modified since the package was installed,

we assume that the new key is equally trustworthy, and import it
automatically.