The Great Denaming: remove 'fedup' and 'fedora' where possible
It has been suggested that some users might be dubious of a tool named
'fedup', for reasons I can't begin to fathom.

This commit removes explicit references to the strings 'fedup' or
'fedora' wherever possible; where needed, the variable __package__ is
used instead (which is still 'fedup', but that could be changed by
installing to a different path.)

It also changes all the imports to be relative imports, which is
probably helpful for testing stuff from a local checkout, so that's cool
too.