fedup.treeinfo: dubious hack to fix unittests
When we import treeinfo using unittest discovery, __package__ is None,
which causes an error when we try to concatenate that with '.treeinfo'.

There might be a smarter way to make __package__ have a reasonable value
(or to set our logger name), but for now this will make the unittests
work.