fetch/verify .treeinfo.signed if gpgcheck is on (#877623)
If instrepo.gpgcheck is True, we will try to fetch .treeinfo.signed and
verify its GPG signature; if that fails, raise an exception and bail
out.

Since we also use .treeinfo to check the integrity of the boot images,
if the .treeinfo signature is good then we can assume that the boot
images are also authentic.