Add/import keys for instrepo so we can check .treeinfo.signed
In order to check .treeinfo.signed we need to find its key. In the
absence of a user-supplied key (which will be added in a future patch),
we need a default key - just like the default repo.

Once we have a key file, we import it into fedup's personal keyring.
We can then use that in check_signed_file to verify .treeinfo.signed, as
expected.