drop extra boot args (#964303)
We have a systemd generator that starts the upgrade, so we don't need to
put 'systemd.unit=upgrade.unit' in the boot args to make that happen.

As for SELinux stuff, eparis says that if the upgrade breaks because of
incompatibilities between the two policies, that's a bug in the policy
and I should make the SELinux guys fix it. Works for me!

This will fix problems where people run fedup, then install a new
kernel and end up with 'systemd.unit=upgrade.unit' in their boot args.

github: fixes #24