add system-upgrade-generator
Rather than depending on 'systemd.unit=system-upgrade.target', look for
the files that indicate that the system is ready for upgrade:

1. /system-upgrade (symlink to directory)
2. /system-upgrade-root (directory)
3. /run/system-upgrade (symlink) *or* /run/initramfs/upgrade.conf (file)

If those three conditions are met, the system is prepared for upgrade,
so (for this boot only) make 'system-upgrade.target' the default taget.

Note that /run/initramfs/upgrade.conf isn't part of the SystemUpgrade
spec, but it's retained here for compatibility with existing Fedora
releases.