Use correct kernel for Xen guests (#1023618)
Xen guests are super-special and need a special kernel to boot.

That kernel exists and it's listed in .treeinfo - you just need to use
the [images-xen] section if you're in a Xen system.

As for Xen hosts, we can use the xen kernel there too - it ain't gonna
hurt nothin'.

So: if we're running under Xen, use the xen kernel.