fedup.download: add interrupt_callback
This adds an interrupt_handler, like yum's, that should allow you to
cancel a single download with CTRL-C, or exit the program by hitting it
twice within two seconds.

You might notice that this doesn't seem to work in Fedora 19;
hilariously, it doesn't seem to work in yum, either, so at least we're
matching the current yum behavior.