Fix ValueError with --addrepo/--repo REPOID (#1045090)
We need to be sure that the argument to --addrepo is actually of the
form "REPOID=[@]URL", or we end up with a ValueError when we try to
split it into two parts later on. So let's check that early.

Also, the --repo argument was deprecated and hidden, and it's now
causing confusion with --instrepo, so drop it.