Fix traceback with '--iso <UNMOUNTABLE>' (#1044083)
There's no 'message' function in fedup.commandline.

So: make device_setup return the mountpoint of the device - unless the
mount fails, in which case it returns None. In that case we can print
the message and exit cleanly from the main script, like most other
errors.

Bonus: remove a gratuitous umount attempt (it happens automatically
thanks to atexit)