Fix "no update for PKG" message for installonly pkgs
Normally we figure out what packages are being updated by looking at the
"remove" items in the RPM transaction: everything that has an update
will have a corresponding "remove" item.

Except if we're installing a new package alongside the existing one
(e.g. new kernels). The kernel package has an update, but there's no
"remove" item.

To fix that, we can also look at the list of pure "install" items - if
we're going to install a new version of this package, then this package
has an update.