Warn the user if upgrade contains no kernels
Various problems can arise if the upgrade contains no kernels, since we
don't build a new initramfs. (Normally we should get a new kernel with
the upgrade but sometimes there's upgrade path problems.)

So: check the upgrade transaction for any package that matches the list
of kernel package names, and warn the user if there's nothing there.

(Yes, in theory we could look through the rpmdb and force-install a
kernel package from the new distribution, but this violates the "no
special cases for packages" principle.)