Fix sudden exit when resizing terminal (#1044987)
So fedup.terminal installed a SIGWINCH handler so that it could update
the terminal size whenever the terminal was changed, rather than polling
to see if the size changed.

If the SIGWINCH happens when we're in select.epoll.poll(), it raises
IOError with errno=4 (EINTR, interrupted system call). The expected
behavior is to simply resume the poll after handling the interrupt.

Unfortunately yum/urlgrabber are kind of stupid about interrupted system
calls, and raise KeyboardError no matter what IOError was raised, which
makes resizing your terminal window work the same as hitting Ctrl-C.

This patch makes us use the same polling mechanism as urlgrabber. Dumb,
but at least we don't silently quit on window resize. And I do love a
patch that removes a file.