fedup.boot: add Initramfs class and helpers
This adds an Initramfs class that's basically a helper for manipulating
initramfs images.

It also adds find_initramfs(), which finds the initramfs that
corresponds to the given kernel version (either according to
http://freedesktop.org/wiki/Specifications/BootLoaderSpec or the
traditional /boot/initramfs-$KERNEL_VER.img path).

Finally we also add a little helper function, current_initramfs(), which
returns the path to the initramfs for the running kernel.