sysprep: pull config files from host initrd into upgrade.img (#1038413)
There's a whole class of bugs where upgrade.img doesn't set up the
system quite the way the normal initrd does, because it doesn't have the
config files from the host system.

For example: some mdraid arrays won't show up in the expected place
unless you have the host's mdadm.conf (see bug #895805).

The existing code (see commit f937c4a) dealt with this problem by
checking for mdadm.conf, and adding it to the initramfs if it's not
empty.

This approach doesn't scale, though: to add these checks for *every*
file that dracut puts into the initrd would mean writing a *lot* of
code just to get a buggy approximation of dracut's behavior.
Plus it just doesn't make sense to maintain that logic inside fedup.

Instead, here is a simpler (and more robust) approach: just copy the
config files from the existing initrd into upgrade.img. Easy!

So that's what this patch does. It works like this:

1. list the contents of the current initramfs
2. identify config files
3. extract them to a temporary directory
4. append them to upgrade.img

Identifying config files is the tricky part. Right now we grab
*everything* in /etc (except /etc/*-release), thus overwriting any
existing files with the host's version.

It also turns out that we need /usr/share/zoneinfo because, hey, now we
probably have /etc/localtime, so we're gonna set the timezone, so we
need that timezone data.

This will probably need to be refined over time, but it's a damn sight
better than what we had.