Add --product=XXX flag for upgrades to F21+
Since Fedora 21 has various Products - where F20 and earlier didn't -
the user might want to choose which product-environment to apply to
their system during this upgrade.

The '--product=PRODUCT' flag is really just a shortcut to install the proper
group(s); the assumption is that for any valid $PRODUCT, there will be a
corresponding group named "$PRODUCT-product-environment".