use yum gpgkey dialog if we can't auto-trust key
If we can't automatically import a GPG key, we should at least ask the
user if they trust that key - this matches yum's behavior.

It also keeps us from bailing out of the upgrade when we're trying to
check updates from a repo that has not yet been used (v. common with,
say, rpmfusion-nonfree).

Resolves: rhbz#1066044