cli: add check_state, implement download()
download() is basically skeletal at the moment but it shows what should
be done there - just need to implement the actual Downloader.

check_state() checks the existing state against the commandline
arguments and exits with an error if they're incompatible - most notably
if you try to run 'fedup download' after an in-progress upgrade has been
interrupted, it'll tell you to either --continue or --abort the upgrade.

Also sprinkled some asserts here and there for safety and added the
fallback exception logger.