state: persistent items, State.packagelist
Added the 'persist' section, whose items don't get removed on clear().

This is useful so we can clear download state without losing track of
the data in our datadir - so we can 'fedup clean' the data separately,
and so if you try to change datadir we can move/remove the old data.

Also add the 'packagelist' property, which handles the "packages.list"
file for us.

Also, clarify the status summary a bit.